+++ mountpoint-s3/scripts/fs_latency_bench.sh
@@ -68,6 +68,7 @@ do
         --allow-delete \
         --log-directory=$log_dir \
         --prefix=${S3_BUCKET_TEST_PREFIX} \
+        --log-metrics \
         ${optional_args}
     mount_status=$?
     if [ $mount_status -ne 0 ]; then
@@ -120,7 +121,7 @@ do
 
     # cleanup mount directory
     rm -rf ${mount_dir}
-    
+
     # increase directory size
     dir_size=$(awk "BEGIN {print $dir_size*10}")
 done
@@ -144,6 +145,7 @@ for job_file in "${jobs_dir}"/*.fio; do
     --allow-overwrite \
     --log-directory=$log_dir \
     --prefix=${S3_BUCKET_TEST_PREFIX} \
+    --log-metrics \
     ${optional_args}
   mount_status=$?
   if [ $mount_status -ne 0 ]; then